+++ src/types/form.ts
@@ -1,0 +1,7 @@
+interface ObjectValue { value: string; label: string }
+type Value = string | ObjectValue;
+
+export type {
+    ObjectValue,
+    Value,
+};