+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [0.9.0] - 2022-02-11
+### Added
+- choice component
+- tabs selector component
+
 ## [0.8.0] - 2022-01-27
 ### Added
 - modal component